Refactored filestructure
