+++ src/components/basic-components/info-area/InfoArea.tsx
