+++ src/components/basic-components/info-area/info-area.module.css
