+++ src/components/code-components/copy-area/CopyArea.tsx
